Wrote mutable model protocol and support methods.
A mutable model is one that corresponds to an (immutable) model, but that has
`MutableProperty`s in place of all non-identifier properties. They can be
changed over time, and each property can be independently modified for its
changes. All mutable models implement an `apply(model:)` method that updates
the necessary properties to match the state of the model applied to it. It is
intended that each property of the mutable model will be compared using `==`
when applying, but this implementation is left up to the mutable models
themselves.

As a proof of concept, `MutableStation` is implemented...I need to do the
others next.