Model layer supports stub models, and MutableModel layer talks to Shark.
To solve the problem of dealing with stub associated objects, I've made every
model capable of containing *just* its identifier. This is simple to implement
at the Model level. At the MutableModel level, I moved signal-producing code
from the view controllers in. MutableModels can be created as stubs, but will
automatically connect to Shark and fill out as many of their properties as they
can, while being subscribed to future changes.

This means that instead of view controllers having to explicitly subscribe to
information for a particular entity in the system, all they have to do is get a
MutableModel for that entity, and the Shark communication will happen beneath
the surface (get it?).

Since the view controller is required to display connection-related error
messages, and since some Shark events on an entity's topic may not have to do
with its properties, delegate methods have been implemented on MutableModels.
`receivedError` passes a connection-closing fatal error to the delegate,
which usually means the view will need to be reloaded. `receivedTopicEvent` is
sent when a topic event is received that the MutableModel doesn't know how to
handle.

Ramifications of this change include:

- It's easier to request a ridiculous amount of data from Shark without meaning
  to. Don't map a Collection of Models to MutableModels without good reason.
- We will need to make sure that topic subscriptions are properly closed when a
  MutableModel is released.

The new models compile, but I am still testing that they work as expected.