MutableModels now associate to other MutableModels
Previously (I thought it'd be simpler) properties like MutableVehicle.route and
MutableRoute.stations were static Model types. But this resulted in a lot of
mapping to MutableModel types and managing of separate MutableModel instances
in the places where they were used. Since we no longer auto-subscribe upon the
creation of a MutableModel, it's okay to create them upon access (which still
happens lazily).

In the future, MutableModel.init methods could be memoized to improve
performance.