Made `CanonicalRoute` generic.
This removes coupling between the `CanonicalRoute` computation and various
mutable model types. In the process, the `RouteStop` enum was moved out of
`CanonicalRoute` since generic types cannot nest.

For clarity, and to make future type modifications easier, I typealiased the
station and vehicle types used by `MutableRoute`. This would be nice to do in
the other mutable models to avoid hard-coding associated types in.