Wrap changes to diff table in table update methods.
The problem (identified in testing) was that the *block* in which
`tableView.beginUpdates()` and `tableView.endUpdates()` occurs must modify the
underlying data source for the table. Previously, I was modifying the data
first, and then updating the table. Note that I wasn't able to reproduce this
outside of testing, so it's possible that I am just doing something weird and
this "fix" in unnecessary.