(#3) Replaced PSError with a protocol-oriented error type
PSError unnecessarily followed the NSError pattern from Objective-C Cocoa. It's
a bloated, somewhat complicated class (errors contain dictionaries and can be
nested!) that obscures useful information, e.g. associated information about
*why* an error was thrown.

To address this, I'm moving to a protocol, ProperErrorType, which builds off of
Swift's ErrorType that can be thrown. Any ProperErrorType has a title, message, and
an optional debug message.

Components of the application are expected to implement their own error value
types which conform to ProperErrorType, for example, the connection layer
implements ConnectionError, the model layer ModelError, etc. This is convenient
because it keeps enums of possible errors near the parts of the application
which throw them.

In situations in which a single code path may encounter multiple
ProperErrorTypes, two options are available:

- The caller may be rewritten to accept generic ProperErrorTypes. This is how
  `ProperViewController.displayError` now works.
- The error may be converted to a general-purpose error type, ProperError,
  which itself is a ProperErrorType. This conversion loses the type association
  of the error but preserves is textual description. This commit introduces a
  signal operator, `genericErrors()`, which performs this conversion on
  signals.