(closes #15) Added a cache for meta.last_event calls.
The cache stores (topic, originator) entries, which is sufficient to use to
look up last_event calls. Connection.call now checks the cache and produces a
cache result if one is available.

TopicEvents flowing through Connection.call and Connection.subscribe are added
to the cache if they have an originator (i.e. all channel-based events). The
cache is suggested to store 100 messages, at which point NSCache will begin
evicting.

I did not write a test for the cache, but I confirmed it was working by
starting a producer of a MutableStation twice.