(WIP #5) Display routes collection in a section of the arrivals table.
To accomplish this, the view controller managing the collection must be the
table view controller for the entire arrivals table---we can't use a collection
view controller per usual. To mitigate this, I wrote a view model for the
routes collection, RoutesCollectionViewModel that serves as its data source and
delegate, to keep that logic out of ArrivalsTableViewController. (side note:
this pattern could be applied elsewhere to shrink VCs.)