Refactored RouteBadge to be nib-based, respond to layout changes, and support highlighting
This fixes a bug where the badge would be drawn larger than its container if
the container resized. This was revealed while working on #5. It's especially
problematic in constraint-based layouts, where layout contraints are applied
after the badge is initialized. The key to this fix involves invoking redraw
code in `layoutSubviews`, and ensuring that the bounding rectangles of
sublayers of the badge view (where the actual drawing is done) always match
their parent container.

I took the opportunity to clean up some of the messy UI code in this class. I
removed all implicitly unwrapped optionals, reduced prevalence of redundant
calculations and redraws, and removed some unnecessary abstraction.

The new badge UI is loaded a xib (RouteBadge.xib), which allows us to set
layout constraints on badges the Apple-preferred way (i.e. not
programmatically), and do things like ensure the badge is always a 1:1 size.
The badge view itself now always lives in a container, which takes the shape of
its parent. This means that the view that the badge layers are attached to can
be sized appropriately, regardless of the size its container is asked to take.
The badge label, too, now adjusts its margins based on badge size and
dynamically resizes its text to fit available space and this is set in the xib
and no longer in draw code.

Custom UI is complicated.