Color the navigation bar with route colors (closes #24)
Route color is used to shade the header bars, and it looks pretty nice.
However, the navigation bar is shared throughout the application (it's part of
the NavigationViewController), so changes in appearance persist. To remedy
this, a new `resetNavigationBar` method has been added to ProperViewController,
which resets the bar back to pre-set defaults.