Created xcworkspace with dependencies.
Since switching the dependency model over to Carthage / dynamic libraries, we
didn't have an easy way to view dependency source code. This commit changes
that by using the workspace as a place to import xcodeproj's of dependencies,
which means that they will be indexed by Xcode, and that "jump to definition",
debugging, etc. will Just Work. I went ahead an imported the dependencies that
I directly import, but am avoiding any sub-dependencies...this can easily be
changed later on if we want to peer into what our libraries are doing more
readily.