Use an independent navigation bar for custom styling.
Unlike the implementation of #24, which animates attribute changes on one
navigation bar shared by the application, this commit handles the same problem
by creating an independent navigation bar for view controller scenes that want
to customize the bar (in this case, RouteViewController).

The benefits of this approach are that animation works smoothly and as expected
(no awkward transitions), and that there's less of a risk of a view controller
screwing up the global navigation bar by not unsetting some changes. The
challenges of this approach are that we lose all of the stuff Proper gets "for
free" from Apple's UINavigationController when we're in a view with a custom
navigation bar. We need to manually wire the "back" button to move back (which
this commit does), and we need to manually configure an edge pan gesture
recognizer and a transition animation to get the same back-on-edge-swipe
behavior (which this commit does not do).

I'm committing this so that if some future Proper *needs* to use multiple
independent navigation bars, we can draw from these initial experiments.