Defer removal of LastEventCache items
Instead of removing items from the cache as soon as `void(topic:)` is called,
this change causes the cache removal to be deferred to the end of the current
iteration of the app's run loop. This allows mutable objects / producers to be
interrupted and recreated in the same event loop, i.e. in response to a list of
mutable objects changing.