Genericized the StationUpcomingTableViewCell into StationUpcomingCell protocol.
We want to be able to use the "upcoming station" view in HeaderFooterViews as
well as TableViewCells. Abstracting to this protocol allows us to re-use the UI
code in both classes. This commit introduces TableViewCell and HeaderFooterView
classes that implement the protocol, and refactors the upcoming station xib to
contain a generic view that is used in both classes.