Color station badges using a pseudorandom seed.
Station colors are now randomly assigned, but done so with a Mersenne Twister
generator seeded off from the station identifier's hash value. This means that
stations *should* always be given the same color, which aids in
recognizability. Hash values are allowed to change across runs of the
application, so occasional color switches are possible.