Location management and tracking on POI view.
This commit introduces `Location`, which creates location signal producers and
acts as a delegate for CLLocationManager. `Location` is currently wired into
POIViewController, and the map is configured to adjust as the user's location
changes.

TODO: the POI view will need to be able to be used with the user's location
*or* with a station's `position`. Thankfully, this should be quite easy to do
since either source has the same signal type.