Static and device-location based points in the POI view.
In the view controller, there are now two producers, `staticLocation` and
`deviceLocation`. By default, the view starts `deviceLocation` and tracks the
user's location, but if a value is sent along `staticLocation`, the view
changes to represent it instead. This allows the view to be used for current
location ("show stations near me") or for another point of interest ("show
stations near x").