Store station and route data with arrival times.
The (ETA, ETD) messages returned from Timetable are contextless without the
route and station they correspond to. To make it easier to store (station,
route, arrival time) data, the `Arrival` struct was renamed to `ArrivalTime`,
and a new `Arrival` struct which contains a station, route, and ArrivalTime is
in its place.

Arrivals are now sortable. Sorting uses the `eta` of the arrival time.