View model implementation of arrivals.
Lots of refactoring done on POI* and NearbyStationsViewModel. In essence, what
has changed is that the view model's representative data has gone from
[Station] to [Station, [Arrival]]. This has merited some change in the way the
view model and its consumers work. Now, the view model has a central `model`
property:

  let model: MutableProperty<[(s: Station, a: [Arrival])]>

This `model` is what the UI is bound to, and is what view controllers must
modify while updating the table. The view model's `producer` produces
replacement values for this property; all a consuming view controller has to do
is modify the property with a producer value when the UI is ready to be updated.