Add support for Timetable route-agnostic endpoints.
I've removed the hack to get arrivals for all routes at a station: this
now uses the Timetable rpc directly. This means that the route on a changed
arrival comes from the ID passed by Timetable. We may need to either map
this to known routes for the station in Proper, or construct a
MutableRoute out of it to get pertinent information like name, color,
etc.