Timetable arrival data in the UI.
Currently, arrival data is static. Next challenge is to keep the
arrivals table up to date as arrivals pass.

Temporarily, ArrivalsTableViewController (which is used in the station
view) has been disabled. It needs to be integrated with Timetable, and
needs many of the same behaviors as are in NearbyStationsViewModel. I
think we should refactor the [Station] -> [Arrival] behavior out of
NearbyStationsViewModel into a more generic view model for managing
arrivals on stations. NSVM can be a subclass of this view model, and the
station view can use it.