Remove MutableModelDelegate.
I've been meaning to do this for awhile. Basically, the delegates aren't
being used at all, and they add binding between a MutableModel and a
controller, which gets awkward. The functionality of the delegates have
already been supplanted:

- Receiving TopicEvents from a model's producer: The `producer` property
is a signal which sends all events, regardless of whether they are
handled. Any signal created will get all events and can do with them as
pleased, however, it cannot be determined whether the model did anything
to handle that event.

- Responding to errors: The same `produer` property will get errors
associated to the model's connection. Other models have been handled at
the controller-level with `displayError`.