Rewrite of the (POI) view model and data source system.
This commit changes the way the POI view works, and specifies design
regarding view models, complex signal chains, and data sources in
general. To begin:

I. POINT-OF-INTEREST VIEW MODEL

The POI view model, at a high level, no longer builds a data
representation of the arrivals table. Previously, the final stages of
the POI signal chain were responsible for constructing a table object of
type:

  [(MutableStation, [Arrival])]

This table was reconstructed each time a station search was performed or
an arrival changed. It was passed to the view controller, which
performed a diff on the new table to figure out how to update the UI.

The limitations of this approach became clear as I started building
Arrival lifecycle. Each time the state of an arrival changed (from
upcoming, to due, to departed, etc.), that change would have to rebuild
the entire table and send it onward. This meant that in order to send a
state change, each Arrival would have to know:

- its neighboring Arrivals, so that it could reconstruct an `[Arrival]`
  list containing its new state
- its MutableStation, so that it could form a new table entry

Since signal chains cannot backtrack, this revealed a fundamental
weakness of the previous design: changes to the POI table could only
initiate from one source: the POI view's tracked `Point` (e.g. the
device's location).

II. NEW DESIGN

The POI view model is now designed to tell the controller _what to do_
with the table, rather than what the table _is_. The signal chain sends
a serial stream of _operations_ to the view controller, defined as:

  enum Op {
    case addStation(MutableStation, index: Int)
    case addArrival(Arrival, to: MutableStation)
    case deleteStation(MutableStation)
    case deleteArrival(Arrival, from: MutableStation)
    case reorderStation(MutableStation, index: Int)
  }

The controller, then, responds to these operations by mutating a data model. To
abstract away the internal storage and the bookkeeping needed to keep
track of station positions within the table, the view controller uses a
new data source class (POITableDataSource).

The POI signal chain, recognizing some degree of separation of concerns,
has been split into three view models:

- NearbyStationsViewModel, which produces a dictionary of stations and
their relative distances given a point and search radius
- ArrivalsViewModel, which produces stream of arrivals for a station
given a list of stations
- POIViewModel, which combines signals from the above two view models

The overall signal chain, from the input of the controller, can be visualized:

		  |
		  | (Point, SearchRadius)
		  |
	+---------v--------+         +-----------------------+
	|                  |         |                       |
	|  NearbyStations  +--------->  Arrivals (timetable) |
	|                  |         |                       |
	+--------------+---+         +----+------------------+
		       |                  |
		       |                  |
  [Station: Distance]  |     +-------+    |   (Station, Arrival, Lifecycle)
		       |     |       |    |
		       +----->  POI  <----+
			     |       |
			     +---+---+
				 |
				 | Op
				 v

Each box represents a view model, arrows are signals.

On the MapKit side, the same table operations are used to create and
remove map annotations as needed.

III. TABLE VIEWS, VIEW MODELS, AND DATA SOURCES

This view is particularly complex for Proper: it involves a multi-input
signal chain and a decent amount of state-tracking, despite this
commit's efforts. I found that complex table views like this one benefit by
separating logic as follows:

1. View models are static, stateless objects that transform data.

A view model's interface is a collection of static functions. These
functions are called by view controllers to transform values, such as
transforming a point to a list of nearby stations, or transforming a
time to its string representation. One view controller may have many
view models, as they exist simply to abstract logic away from the
controller. For view models which form a signal chain, I've introduced a
protocol `SignalChain` which provides a standard interface for chaining
view models together.

2. Data sources manage the table's state.

A view controller holds an instance of a data source, which it uses to
populate the table view. The view controller mutates data in the data
source via its interface, but the data source doesn't ever mutate itself, else
it would diverge from the UI without the view controller knowing about
it. In contrast with view models, which are largely functional and
immutable, data sources perform imperative operations on mutable state.

3. View controllers coordinate the UI by moving data through view models
and into the data source.

"Hot" signals that are actively producing values are bound to a view
controller, where the are injected with side effects that update the
data source and the table's UI. The view controller sends its inputs
(e.g. parent data or sensory data) through view models. It represents
a sort of "declarative/imperative border" in the code where declarative
signal code becomes stateful operations influencing a view. View
controllers should avoid excessive business logic and focus on tying
together supporting objects in the application.