Manipulate the POI table with lists of operations at once.
This smooths animations and reduces CPU time required by minimizing the
number of table view animation blocks needed to add to the table. This
is particularly useful when the list of stations changes, since those
operations are atomic.

This change also interrupts Timetable when a station it's looking up
arrivals for disappears, which prevents redundant data retrieval and crashes in
the data source. Sending the `interrupted` event propagates down to
Timetable and the call is cancelled.