Prevent the continuation from being immediately cancelled.
Originally, the signal __originating from__ the first call to Timetable is
passed into the view model. This means that when that first call is
completed (when the caller finishes receiving the response) the signal
gets a `completed` event and is shut down. This was causing calls to
`more` to be immediately discarded, since they were being performed on a
dead signal.

This change prevents the call signal from forwarding its `completed`
event, thus keeping the signal from `Timetable.visits` alive until a
downstream consumer terminates it, allowing for `more` calls to pass
along.