Connection restarts when interrupted, and shows connection message in-app.
Introduces a new SignalProducer operator, `restarting`, which restarts the
producer after a delay on failure.

Implementing this required writing a new top-level view controller,
ToastNotificationViewController, which sits between the
UINavigationViewController and the rest of the VC hierarchy. Because of
this...and because I didn't want to figure out how to make it fit within the
storyboard, I've ditched the storyboard in favor of programmatically
constructing the UI and VC hierarchy.

Added an outer toast notification VC. Removed storyboarding in the process :-)

Use Config's makeConnection method

Nits and minor cleanup

Remove storyboard and update xcodeproj

Restarting signal operator and its usage in Connection

Always show the oldest toast notification

This helps in instances where, for instance, the connection drops and causes a
call to fail. The connection dropping will toast "Connection interrupted.
Reconnecting..." while the call will toast about its own failure. The first
toast is the most informative, since it speaks to the root cause of the error.